requirements.txt

# ===== Torch CPU Wheels (for Spaces) =====
--find-links https://storage.googleapis.com/torch-cpu/torch_stable.html
--extra-index-url https://download.pytorch.org/whl/cpu

# ===== Core Scientific Stack =====
numpy          # compatible with most PyTorch CPU wheels
scipy             # optional but stable with numpy 1.26

# ===== PyTorch =====
torch
torchvision

# ===== App / ML Libraries =====
gradio
fastapi
uvicorn
pydantic
huggingface_hub
scikit-learn
sentence-transformers
transformers

# ===== PDF / Document Processing =====
pypdf
PyMuPDF
python-docx

# ===== Visualization & Word Analysis =====
wordcloud
matplotlib

# ===== (Optional utilities if needed) =====
# If you use caching / async inference, you can also add:
# aiohttp
# tqdm
